Full rewrite + new features
I had to add support for the OAuth2 discovery document, and I was doing
this I realized that the existing structure didn't work well enough for
it.

So, I started making changes but this quickly escalated to a full
BC-breaking rewrite.

Highlights:

* This library is now becoming a general OAuth2 client library, with the
  fetch wrapper being just 1 of its features.
* Now requires Node 16 due to the use of `btoa()`.
* Better, declarative setup. Fewer crazy options objects.

I'm making no attempt to preserve backwards compatibility (sorry!) but I
think migrating should not be incredibly difficult. All the same
features are there, the setup is just different.

This library may need a new name!